requirements.txt

theano
cython
h5py
lasagne
numpy
scipy
nose
pillow
pyglet
